+++ book_manager/src/book_manager/entities/editorial.py
@@ -1,0 +1,5 @@
+from .entidad_base import EntidadConNombre
+
+
+class Editorial(EntidadConNombre):
+    """Editorial o distribuidora que le provee los libros a la librería."""