+++ book_manager/src/book_manager/entities/tipo_cotizacion.py
@@ -1,0 +1,5 @@
+from .entidad_base import EntidadConNombre
+
+
+class TipoCotizacion(EntidadConNombre):
+    """Oficial, Blue, MEP, etc."""